+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+### Fixed
+
+## [0.48.0] - 2026-02-17
+
+### Added
+
 - Official support for Python 3.14 (include Python 3.14 in unit test matrix on GitHub Actions) ([#801](https://github.com/Open-EO/openeo-python-client/issues/801))
 - Add `on_response_headers_sync` option to  `openeo.connect`/`Connection` to set default `on_response_headers` handler for sync requests at connection level.
 
@@ -16,8 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Harden STAC parsing of "bands"/"eo:bands" when "name" field is not present (which is technically not a required field) ([#860](https://github.com/Open-EO/openeo-python-client/issues/860))
 - `openeo.rest.conversions.timeseries_json_to_pandas`: allow timeseries to be given as a raw JSON dump or path to JSON file.
-
-### Removed
 
 ### Fixed